Cleaned up build script, added headings to include file.
The build script was cleaned up to be more consistent, and to use the low level
SCR_INSALL script instead of building the .H file directly.  This should now be
a good example and template of the build script for other libraries.

Added comments in STUFF.INS.PAS to divide the data type section according to
related functions.